refactor: move shared domain types from agg/types.ts to lib/types.ts
QuestionType, Question, Cell, NaStats, Slice, Axis, Tab are used
across components, export formatters, and lib modules — they belong
in the shared types file rather than the agg-internal module.
agg/types.ts retains Shape, TabData, calcPct and re-exports the
moved types for backward compatibility within agg/.
