+++ src/components/aggregation/ChartCardBody.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { onMount } from "svelte";
-  import type { Tab } from "../../lib/agg/types";
+  import type { Tab } from "../../lib/types";
   import { Chart, getSeriesColor, getThemeColors, type PaletteId } from "../../lib/chartConfig";
   import type { ChartType } from "./viewTypes";
   import type { ChartConfiguration } from "chart.js";
