+++ src/components/aggregation/CrossTable.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import type { Tab, Slice } from "../../lib/agg/types";
+  import type { Tab, Slice } from "../../lib/types";
   import { t } from "../../lib/i18n.svelte";
   import type { Basis } from "./viewTypes";
   import { formatN } from "../../lib/format";
