+++ src/lib/agg/types.ts
@@ -1,25 +1,13 @@
-/** New aggregation type system */
+/** Aggregation-internal types */
 
-export type QuestionType = "SA" | "MA" | "NA";
+export type { QuestionType, Question, Cell, NaStats, Slice, Axis, Tab } from "../types";
 
-export interface Question {
-  type: QuestionType;
-  code: string;
-  columns: string[];
-  codes: string[];
-  label: string;
-  labels: Record<string, string>;
-}
+import type { Question, Tab } from "../types";
 
 export type Shape = Pick<Question, "type" | "columns" | "codes">;
 
-export type Axis = Pick<Question, "code" | "label" | "labels">;
-
-export interface Cell {
-  count: number;
-  /** 百分率。n=0（有効回答なし）のとき null（算出不能） */
-  pct: number | null;
-}
+/** aggregate() の戻り値 — rawdata 由来のみ */
+export type TabData = Pick<Tab, "codes" | "slices">;
 
 /** n=0（有効回答なし）のとき null（算出不能）を返す */
 export function calcPct(count: number, n: number): number | null {
@@ -41,34 +29,4 @@ if (import.meta.vitest) {
   test("calcPct: count=0, n>0 → 0", () => {
     expect(calcPct(0, 10)).toBe(0);
   });
-}
-
-export interface Slice {
-  code: string | null;
-  n: number;
-  cells: Cell[];
-  stats?: NaStats;
-}
-
-/** aggregate() の戻り値 — rawdata 由来のみ */
-export type TabData = Pick<Tab, "codes" | "slices">;
-
-export interface NaStats {
-  n: number;
-  mean: number | null;
-  median: number | null;
-  sd: number | null;
-  min: number | null;
-  max: number | null;
-}
-
-/** 消費用フラット型 */
-export interface Tab {
-  type: QuestionType;
-  questionCode: string;
-  label: string;
-  by: Axis | null;
-  codes: string[];
-  labels: Record<string, string>;
-  slices: Slice[];
 }
